+++ src/main/kotlin/io/kjson/JSONLinesPipeline.kt
@@ -64,13 +64,13 @@ class JSONLinesPipeline<R>(
             State.BOM_POSSIBLE -> {
                 state = State.INITIAL
                 if (ch != BOM && !isSpaceCharacter(ch)) {
-                    child = getAssembler(ch, parseOptions, "/$count")
+                    child = getAssembler(ch, parseOptions, "/$count", 0)
                     state = State.CHILD
                 }
             }
             State.INITIAL -> {
                 if (!isSpaceCharacter(ch)) {
-                    child = getAssembler(ch, parseOptions, "/$count")
+                    child = getAssembler(ch, parseOptions, "/$count", 0)
                     state = State.CHILD
                 }
             }